Feature/update param logic and validation

+++ git_branch.bat
@@ -1,5 +1,5 @@
 echo off
-
+git fetch
 rem validate input
 if [%1] == [] goto usage
 
@@ -14,24 +14,47 @@ if NOT [%ResultUncommittedChanges%]==[] goto uncommitted
 rem set the variable ResultRemoteBranchExists to empty string
 set ResultRemoteBranchExists=
 set i=
-for /f %%i in ('git branch -l *%~1/%2*') do set ResultRemoteBranchExists=%%i
+for /f %%i in ('git branch -r -l *%~1/%2*') do set ResultRemoteBranchExists=%%i
 if [%ResultRemoteBranchExists%]==[] goto  notFound
 
-git fetch
+rem - check if the users specified 2 or 3 params
+if [%3] == [] goto TwoParam
+
+
 git checkout %~1/%2
 git fetch origin 
 git reset --hard origin/%~1/%2
 rundll32 user32.dll,MessageBeep
 
 
-git checkout -b users/pearj01/%1/%2/%3
+git checkout -b users/%USERNAME%/%1/%2/%3
 
 
-exit /B 1
+exit /B 0
 
 echo on
 
+
+:TwoParam
+rem - for when use specifies remote as single parameter
+git fetch
+git checkout %~1
+git fetch origin 
+git reset --hard origin/%~1
+rundll32 user32.dll,MessageBeep
+
+
+git checkout -b users/%USERNAME%/%1/%2
+
+
+exit /B 0
+
+echo on
+
+
+
 :usage
+rem - help file
 @echo -----------------------------------------------------------------------------
 @echo %0 creates a new feature branch
 @echo     - checks if working tree is clean
@@ -49,6 +72,7 @@ echo on
 
 
 :uncommitted
+rem - ERROR - uncommitted changes in current working tree
 @echo -------------------------------------------------------------------------
 @echo error uncommitted changes!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 @echo *******exiting script before we lose anything important*******
@@ -58,6 +82,7 @@ exit /b 1
 echo on
 
 :notFound
+rem - ERROR - remote branch not found
 @echo -------------------------------------------------------------------------
 @echo remote branch not found  %1/%2
 @echo *******exiting script*******
